+++ src/Concerns/Eventable.php
@@ -9,7 +9,7 @@ use Illuminate\Database\Eloquent\Relations\MorphMany;
 /**
  * @method MorphMany morphMany(string $related, string $name, ?string $type = null, ?string $id = null, ?string $localKey = null)
  */
-trait HasEvents
+trait Eventable
 {
     public function addEvent(BackedEnum $event, mixed $data = null): Event
     {